+++ src/main/java/madsciencemod/cards/Booster.java
@@ -17,7 +17,7 @@ public class Booster extends AbstractMadScienceCard {
     public static final String DESCRIPTION = cardStrings.DESCRIPTION;
     private static final int COST = 0;
     private static final int FUEL_AMT = 2;
-    private static final int UPGRADE_FUEL_AMT = 2;
+    private static final int UPGRADE_FUEL_AMT = 1;
     private static final int DRAW_AMT = 2;
     private static final CardType TYPE = CardType.SKILL;
     private static final CardRarity RARITY = CardRarity.RARE;