+++ src/main/java/madsciencemod/cards/Reconstitute.java
@@ -8,6 +8,8 @@ import com.megacrit.cardcrawl.dungeons.AbstractDungeon;
 import com.megacrit.cardcrawl.localization.CardStrings;
 import com.megacrit.cardcrawl.monsters.AbstractMonster;
 
+import madsciencemod.actions.common.GainFuelAction;
+import madsciencemod.powers.FuelPower;
 import madsciencemod.powers.ReconstitutePower;
 
 public class Reconstitute extends AbstractMadScienceCard {
@@ -15,8 +17,9 @@ public class Reconstitute extends AbstractMadScienceCard {
     private static final CardStrings cardStrings = CardCrawlGame.languagePack.getCardStrings(ID);
     public static final String NAME = cardStrings.NAME;
     public static final String DESCRIPTION = cardStrings.DESCRIPTION;
-    private static final int COST = 1;
-    private static final int UPGRADED_COST = 0;
+    public static final String UPGRADE_DESCRIPTION = cardStrings.UPGRADE_DESCRIPTION;
+    public static final String[] EXTENDED_DESCRIPTION = cardStrings.EXTENDED_DESCRIPTION;
+    private static final int COST = 0;
     private static final int MULTIPLIER = 1;
     private static final CardType TYPE = CardType.SKILL;
     private static final CardRarity RARITY = CardRarity.RARE;
@@ -29,7 +32,29 @@ public class Reconstitute extends AbstractMadScienceCard {
 
     @Override
     public void use(AbstractPlayer p, AbstractMonster m) {
-        AbstractDungeon.actionManager.addToBottom(new ApplyPowerAction(p, p, new ReconstitutePower(p, this.magicNumber), this.magicNumber));
+        if (this.upgraded) {
+            int gain = FuelPower.fuelSpentThisTurn * this.magicNumber;
+            AbstractDungeon.actionManager.addToBottom(new GainFuelAction(gain));
+        } else {
+            AbstractDungeon.actionManager.addToBottom(new ApplyPowerAction(p, p, new ReconstitutePower(p, this.magicNumber), this.magicNumber));
+        }
+    }
+
+    @Override
+    public void applyPowers() {
+        if (this.upgraded) {
+            int gain = FuelPower.fuelSpentThisTurn * this.magicNumber;
+            this.rawDescription = UPGRADE_DESCRIPTION + EXTENDED_DESCRIPTION[0] + gain + EXTENDED_DESCRIPTION[1];
+            this.initializeDescription();
+        }
+    }
+
+    @Override
+    public void onMoveToDiscard() {
+        if (this.upgraded) {
+            this.rawDescription = UPGRADE_DESCRIPTION;
+            this.initializeDescription();
+        }
     }
 
     @Override
@@ -41,7 +66,8 @@ public class Reconstitute extends AbstractMadScienceCard {
     public void upgrade() {
         if (!this.upgraded) {
             this.upgradeName();
-            this.upgradeBaseCost(UPGRADED_COST);
+            this.rawDescription = UPGRADE_DESCRIPTION;
+            this.initializeDescription();
         }
     }
 }