+++ src/main/java/madsciencemod/cards/JunkLaser.java
@@ -23,7 +23,6 @@ public class JunkLaser extends AbstractMadScienceCard {
     public static final String[] EXTENDED_DESCRIPTION = cardStrings.EXTENDED_DESCRIPTION;
     private static final int COST = 2;
     private static final int DAMAGE = 14;
-    private static final int UPGRADE_DAMAGE = 2;
     private static final int EXTRA_DAMAGE = 3;
     private static final int UPGRADE_EXTRA_DAMAGE = 1;
     private static final CardType TYPE = CardType.ATTACK;
@@ -88,7 +87,6 @@ public class JunkLaser extends AbstractMadScienceCard {
         if (!this.upgraded) {
             this.upgradeName();
             this.upgradeMagicNumber(UPGRADE_EXTRA_DAMAGE);
-            this.upgradeDamage(UPGRADE_DAMAGE);
         }
     }
 }