+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 * Changed Auto-Strike: now deals 3(5) damage when drawn, instead of 3(4).
 * Changed Booster: now gains 2(3) Fuel instead of 2(4)
 * Changed Flammable Wall: now only gains 16(19) Block or 2(3) Fuel, instead of 15(18) or 3(4) Fuel.
-* Changed Junk Laser: now deals 14(16) damage, +3(4) for every 0 cost card
+* Changed Junk Laser: now deals 14 damage +3(+4) for every 0 cost card instead of +2(+3)
 * Changed Reconstitute: upgrade now instantly gains 1 Fuel for each Fuel spend this turn.
 * Changed Funnel: is now rare instead of common
 * Fixed: Self Destruct now correctly exhausts